Refactor manager order documents panel
